Fix parsing and rendering of the universal selector.
The universal selector was recognized during parsing, but the
character pointer was not advanced, resulting in infinite loop.
Added the missing increment.

When rendering, the universal selector was treated like other
special characters and escaped.  Added special case for it.

The tests didn't test for the universal selector, added few
cases to cover it.